requirements.txt
pinecone
openai
streamlit
python-dotenv

